+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 Releases prior to 7.0 has been removed from this file to declutter search results; see the [archived copy](https://github.com/dipdup-io/dipdup/blob/8.0.0b5/CHANGELOG.md) for the full list.
 
-## [Unreleased]
+## [8.6.2] - 2026-09-23
 
 ### Fixed
 
@@ -861,7 +861,8 @@ Releases prior to 7.0 has been removed from this file to declutter search result
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[Unreleased]: https://github.com/dipdup-io/dipdup/compare/8.6.1...HEAD
+[Unreleased]: https://github.com/dipdup-io/dipdup/compare/8.6.2...HEAD
+[8.6.2]: https://github.com/dipdup-io/dipdup/compare/8.6.1...8.6.2
 [8.6.1]: https://github.com/dipdup-io/dipdup/compare/8.6.0...8.6.1
 [8.6.0]: https://github.com/dipdup-io/dipdup/compare/8.5.2...8.6.0
 [8.5.2]: https://github.com/dipdup-io/dipdup/compare/8.5.1...8.5.2